split monolithic savedcmd_parser.py file into separate files
